+++ source/common/http/codes.h
@@ -74,6 +74,7 @@ public:
   static bool is5xx(uint64_t code) { return code >= 500 && code < 600; }
 
   static bool isGatewayError(uint64_t code) { return code >= 502 && code < 505; }
+  static bool isRateLimited(uint64_t code) { return code == 429; }
 
   static std::string groupStringForResponseCode(Code response_code);
 };